move upgrade/app_test.go to test_helpers.go
